Move common classes
