+++ core/src/main/java/com/scalar/db/common/ActiveTransactionManagedTransactionManager.java
@@ -1,20 +1,18 @@
-package com.scalar.db.transaction.common;
+package com.scalar.db.common;
 
 import com.scalar.db.api.Delete;
+import com.scalar.db.api.DistributedTransaction;
 import com.scalar.db.api.Get;
 import com.scalar.db.api.Mutation;
 import com.scalar.db.api.Put;
 import com.scalar.db.api.Result;
 import com.scalar.db.api.Scan;
-import com.scalar.db.api.TwoPhaseCommitTransaction;
 import com.scalar.db.config.DatabaseConfig;
 import com.scalar.db.exception.transaction.CommitException;
 import com.scalar.db.exception.transaction.CrudException;
-import com.scalar.db.exception.transaction.PreparationException;
 import com.scalar.db.exception.transaction.RollbackException;
 import com.scalar.db.exception.transaction.TransactionException;
 import com.scalar.db.exception.transaction.UnknownTransactionStatusException;
-import com.scalar.db.exception.transaction.ValidationException;
 import com.scalar.db.util.ActiveExpiringMap;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import java.util.List;
@@ -22,17 +20,17 @@ import java.util.Optional;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public abstract class ActiveTransactionManagedTwoPhaseCommitTransactionManager
-    extends AbstractTwoPhaseCommitTransactionManager {
+public abstract class ActiveTransactionManagedTransactionManager
+    extends AbstractDistributedTransactionManager {
 
   private static final long TRANSACTION_EXPIRATION_INTERVAL_MILLIS = 1000;
 
   private static final Logger logger =
-      LoggerFactory.getLogger(AbstractTwoPhaseCommitTransactionManager.class);
+      LoggerFactory.getLogger(ActiveTransactionManagedTransactionManager.class);
 
-  private final ActiveExpiringMap<String, TwoPhaseCommitTransaction> activeTransactions;
+  private final ActiveExpiringMap<String, DistributedTransaction> activeTransactions;
 
-  public ActiveTransactionManagedTwoPhaseCommitTransactionManager(DatabaseConfig config) {
+  public ActiveTransactionManagedTransactionManager(DatabaseConfig config) {
     activeTransactions =
         new ActiveExpiringMap<>(
             config.getActiveTransactionManagementExpirationTimeMillis(),
@@ -47,7 +45,7 @@ public abstract class ActiveTransactionManagedTwoPhaseCommitTransactionManager
             });
   }
 
-  private void add(ActiveTransaction transaction) throws TransactionException {
+  private void add(DistributedTransaction transaction) throws TransactionException {
     if (activeTransactions.putIfAbsent(transaction.getId(), transaction) != null) {
       transaction.rollback();
       throw new IllegalStateException(
@@ -60,7 +58,7 @@ public abstract class ActiveTransactionManagedTwoPhaseCommitTransactionManager
   }
 
   @Override
-  public TwoPhaseCommitTransaction resume(String txId) {
+  public DistributedTransaction resume(String txId) {
     return activeTransactions
         .get(txId)
         .orElseThrow(
@@ -72,18 +70,18 @@ public abstract class ActiveTransactionManagedTwoPhaseCommitTransactionManager
   }
 
   @Override
-  protected TwoPhaseCommitTransaction activate(TwoPhaseCommitTransaction transaction)
+  protected DistributedTransaction activate(DistributedTransaction transaction)
       throws TransactionException {
     return new ActiveTransaction(super.activate(transaction));
   }
 
-  private class ActiveTransaction extends AbstractTwoPhaseCommitTransaction
-      implements WrappedTwoPhaseCommitTransaction {
+  private class ActiveTransaction extends AbstractDistributedTransaction
+      implements WrappedDistributedTransaction {
 
-    private final TwoPhaseCommitTransaction transaction;
+    private final DistributedTransaction transaction;
 
     @SuppressFBWarnings("EI_EXPOSE_REP2")
-    private ActiveTransaction(TwoPhaseCommitTransaction transaction) throws TransactionException {
+    private ActiveTransaction(DistributedTransaction transaction) throws TransactionException {
       this.transaction = transaction;
       add(this);
     }
@@ -129,16 +127,6 @@ public abstract class ActiveTransactionManagedTwoPhaseCommitTransactionManager
     }
 
     @Override
-    public void prepare() throws PreparationException {
-      transaction.prepare();
-    }
-
-    @Override
-    public void validate() throws ValidationException {
-      transaction.validate();
-    }
-
-    @Override
     public void commit() throws CommitException, UnknownTransactionStatusException {
       transaction.commit();
       remove(getId());
@@ -154,9 +142,9 @@ public abstract class ActiveTransactionManagedTwoPhaseCommitTransactionManager
     }
 
     @Override
-    public TwoPhaseCommitTransaction getOriginalTransaction() {
-      if (transaction instanceof WrappedTwoPhaseCommitTransaction) {
-        return ((WrappedTwoPhaseCommitTransaction) transaction).getOriginalTransaction();
+    public DistributedTransaction getOriginalTransaction() {
+      if (transaction instanceof WrappedDistributedTransaction) {
+        return ((WrappedDistributedTransaction) transaction).getOriginalTransaction();
       }
       return transaction;
     }
