+++ core/src/main/java/com/scalar/db/storage/cosmos/SelectStatementHandler.java
@@ -14,10 +14,10 @@ import com.scalar.db.api.ScanAll;
 import com.scalar.db.api.Scanner;
 import com.scalar.db.api.Selection;
 import com.scalar.db.api.TableMetadata;
+import com.scalar.db.common.EmptyScanner;
 import com.scalar.db.common.TableMetadataManager;
 import com.scalar.db.exception.storage.ExecutionException;
 import com.scalar.db.io.Column;
-import com.scalar.db.storage.common.EmptyScanner;
 import com.scalar.db.util.ScalarDbUtils;
 import java.util.ArrayList;
 import java.util.Iterator;
