+++ core/src/main/java/com/scalar/db/storage/dynamo/DynamoOperationChecker.java
@@ -6,10 +6,10 @@ import com.scalar.db.api.Mutation;
 import com.scalar.db.api.Put;
 import com.scalar.db.api.TableMetadata;
 import com.scalar.db.common.TableMetadataManager;
+import com.scalar.db.common.checker.ColumnChecker;
+import com.scalar.db.common.checker.OperationChecker;
 import com.scalar.db.exception.storage.ExecutionException;
 import com.scalar.db.io.DataType;
-import com.scalar.db.storage.common.checker.ColumnChecker;
-import com.scalar.db.storage.common.checker.OperationChecker;
 import javax.annotation.concurrent.ThreadSafe;
 
 @ThreadSafe
