+++ core/src/main/java/com/scalar/db/storage/jdbc/ScannerImpl.java
@@ -2,8 +2,8 @@ package com.scalar.db.storage.jdbc;
 
 import com.scalar.db.api.Result;
 import com.scalar.db.api.Scanner;
+import com.scalar.db.common.ScannerIterator;
 import com.scalar.db.exception.storage.ExecutionException;
-import com.scalar.db.storage.common.ScannerIterator;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
