+++ core/src/main/java/com/scalar/db/storage/multistorage/MultiStorage.java
@@ -13,10 +13,10 @@ import com.scalar.db.api.Put;
 import com.scalar.db.api.Result;
 import com.scalar.db.api.Scan;
 import com.scalar.db.api.Scanner;
+import com.scalar.db.common.AbstractDistributedStorage;
 import com.scalar.db.config.DatabaseConfig;
 import com.scalar.db.exception.storage.ExecutionException;
 import com.scalar.db.service.StorageFactory;
-import com.scalar.db.storage.common.AbstractDistributedStorage;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
