+++ core/src/main/java/com/scalar/db/storage/rpc/ScannerImpl.java
@@ -4,9 +4,9 @@ import com.scalar.db.api.Result;
 import com.scalar.db.api.Scan;
 import com.scalar.db.api.Scanner;
 import com.scalar.db.api.TableMetadata;
+import com.scalar.db.common.ScannerIterator;
 import com.scalar.db.exception.storage.ExecutionException;
 import com.scalar.db.rpc.DistributedStorageGrpc;
-import com.scalar.db.storage.common.ScannerIterator;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Iterator;
