+++ core/src/main/java/com/scalar/db/transaction/jdbc/JdbcTransactionManager.java
@@ -6,17 +6,17 @@ import com.scalar.db.api.DistributedTransaction;
 import com.scalar.db.api.Isolation;
 import com.scalar.db.api.SerializableStrategy;
 import com.scalar.db.api.TransactionState;
+import com.scalar.db.common.ActiveTransactionManagedTransactionManager;
 import com.scalar.db.common.TableMetadataManager;
+import com.scalar.db.common.checker.OperationChecker;
 import com.scalar.db.config.DatabaseConfig;
 import com.scalar.db.exception.transaction.TransactionException;
-import com.scalar.db.storage.common.checker.OperationChecker;
 import com.scalar.db.storage.jdbc.JdbcAdmin;
 import com.scalar.db.storage.jdbc.JdbcConfig;
 import com.scalar.db.storage.jdbc.JdbcService;
 import com.scalar.db.storage.jdbc.JdbcUtils;
 import com.scalar.db.storage.jdbc.RdbEngine;
 import com.scalar.db.storage.jdbc.query.QueryBuilder;
-import com.scalar.db.transaction.common.ActiveTransactionManagedTransactionManager;
 import java.sql.SQLException;
 import java.util.UUID;
 import javax.annotation.concurrent.ThreadSafe;
