+++ core/src/test/java/com/scalar/db/common/AbstractDistributedTransactionManagerTest.java
@@ -1,4 +1,4 @@
-package com.scalar.db.transaction.common;
+package com.scalar.db.common;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -6,16 +6,14 @@ import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.mock;
 
 import com.scalar.db.api.Delete;
+import com.scalar.db.api.DistributedTransaction;
 import com.scalar.db.api.Get;
 import com.scalar.db.api.Mutation;
 import com.scalar.db.api.Put;
 import com.scalar.db.api.Scan;
-import com.scalar.db.api.TwoPhaseCommitTransaction;
 import com.scalar.db.exception.transaction.CommitException;
-import com.scalar.db.exception.transaction.PreparationException;
 import com.scalar.db.exception.transaction.RollbackException;
 import com.scalar.db.exception.transaction.UnknownTransactionStatusException;
-import com.scalar.db.exception.transaction.ValidationException;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
@@ -24,16 +22,16 @@ import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 
-public class AbstractTwoPhaseCommitTransactionManagerTest {
+public class AbstractDistributedTransactionManagerTest {
 
   @SuppressFBWarnings("SIC_INNER_SHOULD_BE_STATIC")
   @SuppressWarnings("ClassCanBeStatic")
   @Nested
   public class StateManagedTransactionTest {
 
-    @Mock private TwoPhaseCommitTransaction wrappedTransaction;
+    @Mock private DistributedTransaction wrappedTransaction;
 
-    private AbstractTwoPhaseCommitTransactionManager.StateManagedTransaction transaction;
+    private AbstractDistributedTransactionManager.StateManagedTransaction transaction;
 
     @BeforeEach
     public void setUp() throws Exception {
@@ -41,7 +39,7 @@ public class AbstractTwoPhaseCommitTransactionManagerTest {
 
       // Arrange
       transaction =
-          new AbstractTwoPhaseCommitTransactionManager.StateManagedTransaction(wrappedTransaction);
+          new AbstractDistributedTransactionManager.StateManagedTransaction(wrappedTransaction);
     }
 
     @Test
@@ -69,7 +67,8 @@ public class AbstractTwoPhaseCommitTransactionManagerTest {
     }
 
     @Test
-    public void crud_AfterPrepare_ShouldThrowIllegalStateException() throws PreparationException {
+    public void crud_AfterCommit_ShouldThrowIllegalStateException()
+        throws CommitException, UnknownTransactionStatusException {
       // Arrange
       Get get = mock(Get.class);
       Scan scan = mock(Scan.class);
@@ -82,7 +81,7 @@ public class AbstractTwoPhaseCommitTransactionManagerTest {
       @SuppressWarnings("unchecked")
       List<Mutation> mutations = (List<Mutation>) mock(List.class);
 
-      transaction.prepare();
+      transaction.commit();
 
       // Act Assert
       assertThatThrownBy(() -> transaction.get(get)).isInstanceOf(IllegalStateException.class);
@@ -127,76 +126,8 @@ public class AbstractTwoPhaseCommitTransactionManagerTest {
     }
 
     @Test
-    public void prepare_ShouldNotThrowAnyException() {
+    public void commit_ShouldNotThrowAnyException() {
       // Arrange
-
-      // Act Assert
-      assertThatCode(() -> transaction.prepare()).doesNotThrowAnyException();
-    }
-
-    @Test
-    public void prepare_Twice_ShouldThrowIllegalStateException() throws PreparationException {
-      // Arrange
-      transaction.prepare();
-
-      // Act Assert
-      assertThatThrownBy(() -> transaction.prepare()).isInstanceOf(IllegalStateException.class);
-    }
-
-    @Test
-    public void validate_AfterPrepare_ShouldNotThrowAnyException() throws PreparationException {
-      // Arrange
-      transaction.prepare();
-
-      // Act Assert
-      assertThatCode(() -> transaction.validate()).doesNotThrowAnyException();
-    }
-
-    @Test
-    public void validate_Twice_ShouldThrowIllegalStateException()
-        throws PreparationException, ValidationException {
-      // Arrange
-      transaction.prepare();
-      transaction.validate();
-
-      // Act Assert
-      assertThatThrownBy(() -> transaction.validate()).isInstanceOf(IllegalStateException.class);
-    }
-
-    @Test
-    public void validate_BeforePrepare_ShouldThrowIllegalStateException() {
-      // Arrange
-
-      // Act Assert
-      assertThatThrownBy(() -> transaction.validate()).isInstanceOf(IllegalStateException.class);
-    }
-
-    @Test
-    public void validate_AfterPrepareFailed_ShouldThrowIllegalStateException()
-        throws PreparationException {
-      // Arrange
-      doThrow(PreparationException.class).when(wrappedTransaction).prepare();
-      assertThatThrownBy(() -> transaction.prepare()).isInstanceOf(PreparationException.class);
-
-      // Act Assert
-      assertThatThrownBy(() -> transaction.validate()).isInstanceOf(IllegalStateException.class);
-    }
-
-    @Test
-    public void commit_AfterPrepare_ShouldNotThrowAnyException() throws PreparationException {
-      // Arrange
-      transaction.prepare();
-
-      // Act Assert
-      assertThatCode(() -> transaction.commit()).doesNotThrowAnyException();
-    }
-
-    @Test
-    public void commit_AfterValidate_ShouldNotThrowAnyException()
-        throws PreparationException, ValidationException {
-      // Arrange
-      transaction.prepare();
-      transaction.validate();
 
       // Act Assert
       assertThatCode(() -> transaction.commit()).doesNotThrowAnyException();
@@ -204,34 +135,9 @@ public class AbstractTwoPhaseCommitTransactionManagerTest {
 
     @Test
     public void commit_Twice_ShouldThrowIllegalStateException()
-        throws PreparationException, CommitException, UnknownTransactionStatusException {
+        throws CommitException, UnknownTransactionStatusException {
       // Arrange
-      transaction.prepare();
       transaction.commit();
-
-      // Act Assert
-      assertThatThrownBy(() -> transaction.commit()).isInstanceOf(IllegalStateException.class);
-    }
-
-    @Test
-    public void commit_AfterPrepareFailed_ShouldThrowIllegalStateException()
-        throws PreparationException {
-      // Arrange
-      doThrow(PreparationException.class).when(wrappedTransaction).prepare();
-      assertThatThrownBy(() -> transaction.prepare()).isInstanceOf(PreparationException.class);
-
-      // Act Assert
-      assertThatThrownBy(() -> transaction.commit()).isInstanceOf(IllegalStateException.class);
-    }
-
-    @Test
-    public void commit_AfterValidateFailed_ShouldThrowIllegalStateException()
-        throws PreparationException, ValidationException {
-      // Arrange
-      doThrow(ValidationException.class).when(wrappedTransaction).validate();
-
-      transaction.prepare();
-      assertThatThrownBy(() -> transaction.validate()).isInstanceOf(ValidationException.class);
 
       // Act Assert
       assertThatThrownBy(() -> transaction.commit()).isInstanceOf(IllegalStateException.class);
@@ -256,11 +162,9 @@ public class AbstractTwoPhaseCommitTransactionManagerTest {
 
     @Test
     public void rollback_AfterCommitFailed_ShouldNotThrowAnyException()
-        throws PreparationException, CommitException, UnknownTransactionStatusException {
+        throws CommitException, UnknownTransactionStatusException {
       // Arrange
       doThrow(CommitException.class).when(wrappedTransaction).commit();
-
-      transaction.prepare();
       assertThatThrownBy(() -> transaction.commit()).isInstanceOf(CommitException.class);
 
       // Act Assert
@@ -268,34 +172,9 @@ public class AbstractTwoPhaseCommitTransactionManagerTest {
     }
 
     @Test
-    public void rollback_AfterPrepareFailed_ShouldNotThrowAnyException()
-        throws PreparationException {
-      // Arrange
-      doThrow(PreparationException.class).when(wrappedTransaction).prepare();
-      assertThatThrownBy(() -> transaction.prepare()).isInstanceOf(PreparationException.class);
-
-      // Act Assert
-      assertThatCode(() -> transaction.rollback()).doesNotThrowAnyException();
-    }
-
-    @Test
-    public void rollback_AfterValidateFailed_ShouldNotThrowAnyException()
-        throws PreparationException, ValidationException {
-      // Arrange
-      doThrow(ValidationException.class).when(wrappedTransaction).validate();
-
-      transaction.prepare();
-      assertThatThrownBy(() -> transaction.validate()).isInstanceOf(ValidationException.class);
-
-      // Act Assert
-      assertThatCode(() -> transaction.rollback()).doesNotThrowAnyException();
-    }
-
-    @Test
     public void rollback_AfterCommit_ShouldThrowIllegalStateException()
-        throws PreparationException, CommitException, UnknownTransactionStatusException {
+        throws CommitException, UnknownTransactionStatusException {
       // Arrange
-      transaction.prepare();
       transaction.commit();
 
       // Act Assert
