+++ core/src/test/java/com/scalar/db/common/AbstractTwoPhaseCommitTransactionManagerTest.java
@@ -1,4 +1,4 @@
-package com.scalar.db.transaction.common;
+package com.scalar.db.common;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -6,14 +6,16 @@ import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.mock;
 
 import com.scalar.db.api.Delete;
-import com.scalar.db.api.DistributedTransaction;
 import com.scalar.db.api.Get;
 import com.scalar.db.api.Mutation;
 import com.scalar.db.api.Put;
 import com.scalar.db.api.Scan;
+import com.scalar.db.api.TwoPhaseCommitTransaction;
 import com.scalar.db.exception.transaction.CommitException;
+import com.scalar.db.exception.transaction.PreparationException;
 import com.scalar.db.exception.transaction.RollbackException;
 import com.scalar.db.exception.transaction.UnknownTransactionStatusException;
+import com.scalar.db.exception.transaction.ValidationException;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
@@ -22,16 +24,16 @@ import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 
-public class AbstractDistributedTransactionManagerTest {
+public class AbstractTwoPhaseCommitTransactionManagerTest {
 
   @SuppressFBWarnings("SIC_INNER_SHOULD_BE_STATIC")
   @SuppressWarnings("ClassCanBeStatic")
   @Nested
   public class StateManagedTransactionTest {
 
-    @Mock private DistributedTransaction wrappedTransaction;
+    @Mock private TwoPhaseCommitTransaction wrappedTransaction;
 
-    private AbstractDistributedTransactionManager.StateManagedTransaction transaction;
+    private AbstractTwoPhaseCommitTransactionManager.StateManagedTransaction transaction;
 
     @BeforeEach
     public void setUp() throws Exception {
@@ -39,7 +41,7 @@ public class AbstractDistributedTransactionManagerTest {
 
       // Arrange
       transaction =
-          new AbstractDistributedTransactionManager.StateManagedTransaction(wrappedTransaction);
+          new AbstractTwoPhaseCommitTransactionManager.StateManagedTransaction(wrappedTransaction);
     }
 
     @Test
@@ -67,8 +69,7 @@ public class AbstractDistributedTransactionManagerTest {
     }
 
     @Test
-    public void crud_AfterCommit_ShouldThrowIllegalStateException()
-        throws CommitException, UnknownTransactionStatusException {
+    public void crud_AfterPrepare_ShouldThrowIllegalStateException() throws PreparationException {
       // Arrange
       Get get = mock(Get.class);
       Scan scan = mock(Scan.class);
@@ -81,7 +82,7 @@ public class AbstractDistributedTransactionManagerTest {
       @SuppressWarnings("unchecked")
       List<Mutation> mutations = (List<Mutation>) mock(List.class);
 
-      transaction.commit();
+      transaction.prepare();
 
       // Act Assert
       assertThatThrownBy(() -> transaction.get(get)).isInstanceOf(IllegalStateException.class);
@@ -126,8 +127,76 @@ public class AbstractDistributedTransactionManagerTest {
     }
 
     @Test
-    public void commit_ShouldNotThrowAnyException() {
+    public void prepare_ShouldNotThrowAnyException() {
       // Arrange
+
+      // Act Assert
+      assertThatCode(() -> transaction.prepare()).doesNotThrowAnyException();
+    }
+
+    @Test
+    public void prepare_Twice_ShouldThrowIllegalStateException() throws PreparationException {
+      // Arrange
+      transaction.prepare();
+
+      // Act Assert
+      assertThatThrownBy(() -> transaction.prepare()).isInstanceOf(IllegalStateException.class);
+    }
+
+    @Test
+    public void validate_AfterPrepare_ShouldNotThrowAnyException() throws PreparationException {
+      // Arrange
+      transaction.prepare();
+
+      // Act Assert
+      assertThatCode(() -> transaction.validate()).doesNotThrowAnyException();
+    }
+
+    @Test
+    public void validate_Twice_ShouldThrowIllegalStateException()
+        throws PreparationException, ValidationException {
+      // Arrange
+      transaction.prepare();
+      transaction.validate();
+
+      // Act Assert
+      assertThatThrownBy(() -> transaction.validate()).isInstanceOf(IllegalStateException.class);
+    }
+
+    @Test
+    public void validate_BeforePrepare_ShouldThrowIllegalStateException() {
+      // Arrange
+
+      // Act Assert
+      assertThatThrownBy(() -> transaction.validate()).isInstanceOf(IllegalStateException.class);
+    }
+
+    @Test
+    public void validate_AfterPrepareFailed_ShouldThrowIllegalStateException()
+        throws PreparationException {
+      // Arrange
+      doThrow(PreparationException.class).when(wrappedTransaction).prepare();
+      assertThatThrownBy(() -> transaction.prepare()).isInstanceOf(PreparationException.class);
+
+      // Act Assert
+      assertThatThrownBy(() -> transaction.validate()).isInstanceOf(IllegalStateException.class);
+    }
+
+    @Test
+    public void commit_AfterPrepare_ShouldNotThrowAnyException() throws PreparationException {
+      // Arrange
+      transaction.prepare();
+
+      // Act Assert
+      assertThatCode(() -> transaction.commit()).doesNotThrowAnyException();
+    }
+
+    @Test
+    public void commit_AfterValidate_ShouldNotThrowAnyException()
+        throws PreparationException, ValidationException {
+      // Arrange
+      transaction.prepare();
+      transaction.validate();
 
       // Act Assert
       assertThatCode(() -> transaction.commit()).doesNotThrowAnyException();
@@ -135,9 +204,34 @@ public class AbstractDistributedTransactionManagerTest {
 
     @Test
     public void commit_Twice_ShouldThrowIllegalStateException()
-        throws CommitException, UnknownTransactionStatusException {
+        throws PreparationException, CommitException, UnknownTransactionStatusException {
       // Arrange
+      transaction.prepare();
       transaction.commit();
+
+      // Act Assert
+      assertThatThrownBy(() -> transaction.commit()).isInstanceOf(IllegalStateException.class);
+    }
+
+    @Test
+    public void commit_AfterPrepareFailed_ShouldThrowIllegalStateException()
+        throws PreparationException {
+      // Arrange
+      doThrow(PreparationException.class).when(wrappedTransaction).prepare();
+      assertThatThrownBy(() -> transaction.prepare()).isInstanceOf(PreparationException.class);
+
+      // Act Assert
+      assertThatThrownBy(() -> transaction.commit()).isInstanceOf(IllegalStateException.class);
+    }
+
+    @Test
+    public void commit_AfterValidateFailed_ShouldThrowIllegalStateException()
+        throws PreparationException, ValidationException {
+      // Arrange
+      doThrow(ValidationException.class).when(wrappedTransaction).validate();
+
+      transaction.prepare();
+      assertThatThrownBy(() -> transaction.validate()).isInstanceOf(ValidationException.class);
 
       // Act Assert
       assertThatThrownBy(() -> transaction.commit()).isInstanceOf(IllegalStateException.class);
@@ -162,9 +256,11 @@ public class AbstractDistributedTransactionManagerTest {
 
     @Test
     public void rollback_AfterCommitFailed_ShouldNotThrowAnyException()
-        throws CommitException, UnknownTransactionStatusException {
+        throws PreparationException, CommitException, UnknownTransactionStatusException {
       // Arrange
       doThrow(CommitException.class).when(wrappedTransaction).commit();
+
+      transaction.prepare();
       assertThatThrownBy(() -> transaction.commit()).isInstanceOf(CommitException.class);
 
       // Act Assert
@@ -172,9 +268,34 @@ public class AbstractDistributedTransactionManagerTest {
     }
 
     @Test
-    public void rollback_AfterCommit_ShouldThrowIllegalStateException()
-        throws CommitException, UnknownTransactionStatusException {
+    public void rollback_AfterPrepareFailed_ShouldNotThrowAnyException()
+        throws PreparationException {
       // Arrange
+      doThrow(PreparationException.class).when(wrappedTransaction).prepare();
+      assertThatThrownBy(() -> transaction.prepare()).isInstanceOf(PreparationException.class);
+
+      // Act Assert
+      assertThatCode(() -> transaction.rollback()).doesNotThrowAnyException();
+    }
+
+    @Test
+    public void rollback_AfterValidateFailed_ShouldNotThrowAnyException()
+        throws PreparationException, ValidationException {
+      // Arrange
+      doThrow(ValidationException.class).when(wrappedTransaction).validate();
+
+      transaction.prepare();
+      assertThatThrownBy(() -> transaction.validate()).isInstanceOf(ValidationException.class);
+
+      // Act Assert
+      assertThatCode(() -> transaction.rollback()).doesNotThrowAnyException();
+    }
+
+    @Test
+    public void rollback_AfterCommit_ShouldThrowIllegalStateException()
+        throws PreparationException, CommitException, UnknownTransactionStatusException {
+      // Arrange
+      transaction.prepare();
       transaction.commit();
 
       // Act Assert
